+++ exporters/otlp/otlplog/otlploghttp/internal/x/observ_test.go
@@ -13,9 +13,6 @@ func TestObservability(t *testing.T) {
 	const key = "OTEL_GO_X_OBSERVABILITY"
 	require.Contains(t, Observability.Keys(), key)
 
-	const altKey = "OTEL_GO_X_SELF_OBSERVABILITY"
-	require.Contains(t, Observability.Keys(), altKey)
-
 	t.Run("100", run(setenv(key, "100"), assertDisabled(Observability)))
 	t.Run("true", run(setenv(key, "true"), assertEnabled(Observability, "true")))
 	t.Run("True", run(setenv(key, "True"), assertEnabled(Observability, "True")))