+++ sdk/trace/internal/observ/batch_span_processor.go
@@ -12,7 +12,7 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/sdk"
-	"go.opentelemetry.io/otel/sdk/trace/internal/x"
+	"go.opentelemetry.io/otel/sdk/internal/x"
 	semconv "go.opentelemetry.io/otel/semconv/v1.37.0"
 	"go.opentelemetry.io/otel/semconv/v1.37.0/otelconv"
 )